requirements.txt
pyyaml~=6.0
requests~=2.31.0
pytest~=8.0.0